pyproject.toml
[project]
name = "exllmgrpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diffusers>=0.34.0",
    "grpcio-reflection>=1.73.1",
    "grpcio-tools>=1.73.1",
    "grpcio>=1.73.1",
    "numpy<2",
    "pillow>=11.3.0",
    "transformers>=4.53.1",
]
